# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Compsole API documentation.\n\n### Authenticating\n\nThere are two methods of authenticating to the Compsole API. **Basic Auth** is used solely for the purpose of the Compsole UI. **Api Key Auth** is used for service accounts to authenticate prior to accessing the REST endpoints.\n\n#### Basic Auth\n\nUse the [`/auth/local/login`](#operations-Auth_API-post_auth_local_login) endpoint below to authenticate from the Compsole UI.\n\n#### Api Key Auth\n\nAPI Key Authentication is more complicated.\n\n1. You must retreive your `api_key` and `api_secret` from the Compsole UI after creating a service account.\n2. Use the [`/rest/token`](#operations-Auth_API-post_rest_login) endpoint to retrieve an API Token to use in requests\n3. Place the API Token into the `Authorization` header like so: `Authorization: Bearer \u003capi token here...\u003e`\n\n##### Refresh Tokens\n\nRefresh tokens can be used to renew a session without re-authenticating. The refresh token should be set in the `refresh-token` cookie already, so you can simply make a `POST` request to `/rest/token/refresh` and receive a new API token from this endpoint.\n",
        "title": "Compsole API",
        "contact": {
            "name": "BradHacker",
            "url": "http://github.com/BradHacker/compsole/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
...
    },
    "basePath": "/api",
    "paths": {
...
        "/rest/competition": {
            "get": {
                "security": [
                    {
                        "ServiceAuth": []
                    }
                ],
                "description": "List all Competitions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service API"
                ],
                "summary": "List all Competitions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/rest.CompetitionModel"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ServiceAuth": []
                    }
                ],
                "description": "Create a Competition",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service API"
                ],
                "summary": "Create a Competition",
                "parameters": [
                    {
                        "description": "The competition to create",
                        "name": "competition",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.CompetitionInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/rest.CompetitionModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
        "/rest/competition/{id}": {
            "get": {
                "security": [
                    {
                        "ServiceAuth": []
                    }
                ],
                "description": "Get a Competition",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service API"
                ],
                "summary": "Get a Competition",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx",
                        "description": "The id of the competition",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.CompetitionModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ServiceAuth": []
                    }
                ],
                "description": "Update a Competition",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service API"
                ],
                "summary": "Update a Competition",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx",
                        "description": "The id of the competition",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The updated competition",
                        "name": "competition",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.CompetitionInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/rest.CompetitionModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ServiceAuth": []
                    }
                ],
                "description": "Delete a Competition",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service API"
                ],
                "summary": "Delete a Competition",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx",
                        "description": "The id of the competition",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIError": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "rest.CompetitionInput": {
            "description": "Used as an input model for creating/updating Competitions",
            "type": "object",
            "required": [
                "competition_to_provider",
                "name"
            ],
            "properties": {
                "competition_to_provider": {
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "name": {
                    "type": "string",
                    "example": "ISTS 'XX"
                }
            }
        },
        "rest.CompetitionModel": {
            "description": "Used for Competition endpoints",
            "type": "object",
            "properties": {
                "competition_to_provider": {
                    "$ref": "#/definitions/rest.ProviderEdge"
                },
                "competition_to_teams": {
                    "description": "Edges",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.TeamEdge"
                    }
                },
                "id": {
                    "description": "Fields",
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "name": {
                    "description": "[REQUIRED] The unique name (aka. slug) for the competition.",
                    "type": "string",
                    "example": "Test Competition"
                }
            }
        },
        "rest.ProviderEdge": {
            "description": "Used for Provider in edges",
            "type": "object",
            "properties": {
                "config": {
                    "description": "[REQUIRED] This is the JSON configuration for the provider.",
                    "type": "string",
                    "example": "{...}"
                },
                "id": {
                    "description": "Fields",
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "name": {
                    "description": "[REQUIRED] The unique name (aka. slug) for the provider.",
                    "type": "string",
                    "example": "Local Openstack"
                },
                "type": {
                    "description": "[REQUIRED] The type of provider this is (must match a registered one in https://github.com/BradHacker/compsole/tree/main/compsole/providers)",
                    "type": "string",
                    "example": "OPENSTACK"
                }
            }
        },
...
        "rest.TeamEdge": {
            "description": "Used for Team in edges",
            "type": "object",
            "properties": {
                "id": {
                    "description": "Fields",
                    "type": "string",
                    "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                },
                "name": {
                    "description": "[OPTIONAL] The display name for the team.",
                    "type": "string",
                    "example": "Team 1"
                },
                "team_number": {
                    "description": "[REQUIRED] The team number.",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ServiceAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "UserAuth": {
            "type": "basic"
        }
    },
...
}